@@ -3,7 +3,7 @@ import './App.css';
 import Navbar from './Components/Navbar/Navbar.jsx';
 import Landing from './Components/Landing/Landing.jsx';
 import Footer from './Components/Footer/Footer.jsx';
-import Wraper from './Components/Wraper/Wraper';
+import Wraper from './Components/Wraper/Wraper.jsx';
 
 import { Routes, Route } from 'react-router-dom';
 
